handwired/cmd60: rename all files and folders to lowercase
