+++ keyboards/handwired/cmd60/cmd60.c
@@ -1,4 +1,4 @@
-#include "CMD60.h"
+#include "cmd60.h"
 
 void matrix_init_kb(void) {
 	// put your keyboard start-up code here
