+++ keyboards/handwired/cmd60/config.h
