+++ keyboards/handwired/cmd60/info.json
