+++ keyboards/handwired/cmd60/rules.mk
